backend/pyproject.toml
[project]
name = "rag-content-ingestion"
version = "0.1.0"
description = "RAG Content Ingestion System for textbook content"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "httpx>=0.24.0",
    "slowapi>=0.1.9",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.24.0",
]